pyproject.toml
[project]
name = "ruffo"
version = "0.1.0"
description = "Ruffo - Bot de tienda de mascotas Animalicha con LangChain y Telegram"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # LangChain & LangGraph
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",

    # Telegram
    "aiogram>=3.4.0",

    # Google APIs
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.120.0",

    # Slack
    "slack-sdk>=3.27.0",

    # HTTP Client
    "httpx>=0.27.0",

    # Web Server (FastAPI)
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",

    # Validation & Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Utilities
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
